chore: remove unused error
